compat: keep single stress dir; serialize via cpu:96 + trim log/kv load
Collapse the stress_heavy split back into one stress/ folder to avoid
folder sprawl. Prevent the msan OOM (parallel ~150 GB log/kv chunks
exhausting the 283 GB runner) by serializing the suite with cpu:96 so at
most one heavy chunk runs at a time (REQUIREMENTS(ram) is ignored locally).

Reduce log/kv data footprint (test_log timeout_scale 60->30, test_kv mixed
--seconds 180->90) so a follow-up msan run can measure the new per-chunk
peak; once it drops enough we can lower cpu back to allow parallelism.

Document the compatibility folder layout and placement rule in README so it
is clear which directory a new test belongs to and when a new one is
justified (only for a distinct resource profile, since REQUIREMENTS live per
ya.make/dir).
