+++ ydb/tests/compatibility/README.md
@@ -8,9 +8,23 @@
 
 ## Как написать тест
 
-1. Выбрать подходящую [фикстуру](#фикстуры)
-2. Сделать новый файл с тестом (либо же выбрать имеющийся, если он подходит по области)
-3. Закодировать тест. Как первый шаг, можно скопировать пример из [common/test_example.py](common/test_example.py)
+1. Выбрать [папку](#куда-положить-тест-структура-папок) по функциональной области
+2. Выбрать подходящую [фикстуру](#фикстуры)
+3. Сделать новый файл с тестом (либо же выбрать имеющийся, если он подходит по области)
+4. Закодировать тест. Как первый шаг, можно скопировать пример из [common/test_example.py](common/test_example.py)
+
+## Куда положить тест (структура папок)
+
+Каждая папка — это отдельный `ya.make` (отдельный тестовый модуль). Основное правило: **клади тест в папку по функциональной области фичи, которую проверяешь**.
+
+**Папки по областям** (сюда попадает большинство новых тестов): `indexes`, `olap`, `topics`, `udf`, `s3_backups`, `federated_queries`, `result_set_format`, `streaming`. Внутри папки выбери файл по под-области, а сценарий (mixed / restart / rolling) задаётся выбором [фикстуры](#фикстуры), а НЕ папкой.
+
+**Особые папки** (не по фиче — заведены осознанно):
+- `common` — общие проверки и [пример](common/test_example.py). Старт отсюда, если подходящей области ещё нет.
+- `workloads` — «фича переживает апгрейд»: небольшие ворклоады через **Python ydb SDK** (`self.driver`), фиксированный объём данных, проверка корректности.
+- `stress` — **объёмная нагрузка** через внешний `ydb` CLI (`workload log/kv/tpch`): bulk-upsert, tpch-датасеты. Вынесена отдельно ради **изолированного ресурсного профиля**: тяжёлые чанки едят десятки–сотни ГБ, поэтому у папки высокий `REQUIREMENTS(cpu)`, чтобы локальный планировщик не запускал их параллельно и не ловил OOM.
+
+**Правило про новую папку**: `REQUIREMENTS(cpu/ram)`, `SPLIT_FACTOR`, `SIZE` задаются на `ya.make`, то есть на папку (два `PY3TEST()` в одной директории Arcadia не поддерживает). Заводи новую папку **только** если тесту нужен принципиально иной ресурсный профиль (как у `stress`). В остальных случаях — клади в существующую папку по области.
 
 ## Фикстуры
