--- ydb/tests/compatibility/stress_heavy/ya.make
@@ -1,30 +1,0 @@
-PY3TEST()
-INCLUDE(${ARCADIA_ROOT}/ydb/tests/harness_dep.inc)
-ENV(YDB_CLI_BINARY="ydb/apps/ydb/ydb")
-
-FORK_TEST_FILES()
-FORK_TESTS()
-FORK_SUBTESTS()
-SPLIT_FACTOR(16)
-
-TEST_SRCS(
-    test_stress.py
-)
-
-SIZE(LARGE)
-REQUIREMENTS(ram:32 cpu:96)
-INCLUDE(${ARCADIA_ROOT}/ydb/tests/large.inc)
-
-DEPENDS(
-    ydb/tests/library/compatibility/binaries
-    ydb/apps/ydb
-)
-
-PEERDIR(
-    ydb/tests/library
-    ydb/tests/datashard/lib
-    ydb/tests/stress/simple_queue/workload
-    ydb/tests/library/compatibility
-)
-
-END()
